+++ tests/commands/test_explore.py
@@ -31,7 +31,7 @@ class TestCli(tests.cli.TestCli):
     def test_export_bibtex(self):
         path = tempfile.mktemp()
         result = self.invoke([
-            'lib', 'krishnamurti', 'export', '--bibtex', path
+            'lib', 'krishnamurti', 'export', '--format', 'bibtex', '-o', path
         ])
         self.assertTrue(result.exit_code == 0)
         self.assertTrue(os.path.exists(path))
@@ -51,7 +51,7 @@ class TestCli(tests.cli.TestCli):
     def test_export_yaml(self):
         path = tempfile.mktemp()
         result = self.invoke([
-            'lib', 'krishnamurti', 'export', '--yaml', path
+            'lib', 'krishnamurti', 'export', '--format', 'yaml', '-o', path
         ])
         self.assertTrue(result.exit_code == 0)
         self.assertTrue(os.path.exists(path))
@@ -70,7 +70,7 @@ class TestCli(tests.cli.TestCli):
     def test_citations(self):
         path = tempfile.mktemp()
         result = self.invoke([
-            'lib', 'krishnamurti', 'export', '--yaml', path
+            'lib', 'krishnamurti', 'export', '--format', 'yaml', '-o', path
         ])
         self.assertTrue(result.exit_code == 0)
         self.assertTrue(os.path.exists(path))
@@ -89,7 +89,8 @@ class TestCli(tests.cli.TestCli):
     def test_citations_and_json(self):
         path = tempfile.mktemp()
         result = self.invoke([
-            'citations', 'krishnamurti', 'export', '--json', path
+            'citations', 'krishnamurti', 'export', '--format', 'json', '--out',
+            path
         ])
         self.assertTrue(result.exit_code == 0)
         with open(path) as fd: